Move operators to operators/ directory
