+++ tomviz/operators/CropOperator.cxx
