+++ tomviz/operators/CustomPythonOperatorWidget.cxx
