+++ tomviz/operators/EditOperatorDialog.cxx
