+++ tomviz/operators/EditOperatorDialog.h
