+++ tomviz/operators/EditOperatorWidget.cxx
