+++ tomviz/operators/EditOperatorWidget.h
