+++ tomviz/operators/EditPythonOperatorWidget.ui
