+++ tomviz/operators/Operator.cxx
