+++ tomviz/operators/Operator.h
