+++ tomviz/operators/OperatorDialog.cxx
