+++ tomviz/operators/OperatorFactory.cxx
