+++ tomviz/operators/OperatorFactory.h
