+++ tomviz/operators/OperatorPropertiesPanel.cxx
