+++ tomviz/operators/OperatorPython.cxx
