+++ tomviz/operators/OperatorPython.h
