+++ tomviz/operators/OperatorResult.cxx
