+++ tomviz/operators/OperatorResult.h
