+++ tomviz/operators/OperatorWidget.cxx
