+++ tomviz/operators/SnapshotOperator.cxx
