+++ tomviz/operators/SnapshotOperator.h
